Make exemplar package internal
