+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/view/DefaultAggregation.java
@@ -6,10 +6,10 @@
 package io.opentelemetry.sdk.metrics.view;
 
 import io.opentelemetry.sdk.internal.ThrottlingLogger;
-import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
 import io.opentelemetry.sdk.metrics.internal.aggregator.AggregatorFactory;
 import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
+import io.opentelemetry.sdk.metrics.internal.exemplar.ExemplarFilter;
 import java.util.logging.Level;
 import java.util.logging.Logger;
